Rely on pip to install needed packages
This adds pygments to the list of dependencies, and relies on pip
to install needed packages where possible.

I've left the lxml and pygments instructions for Linux, which are
modifying a system install of Python using sudo, and the OS X
instructions because I don't have a way to test them.