Update WHATWG boilerplate
This uses the computed metadata features to set better defaults across all WHATWG standards, and moves certain scripts (including the service worker registration script) into the boilerplate.

As a drive-by, it removes the "likely out-of-date" for translations as it turns out we have very responsive translators for WHATWG standards and we'd like to be a bit nicer about their work.