Make "bikeshed issues-list" more broadly applicable
"bikshed issues-list" had a number of assumptions built it that made it
impractical to use with anything other than a CSSWG spec. This removes
these assumptions and allows it to be used with any document, while
preserving the existing behavior that defaults to CSSWG conventions,
so that existing uses can stay unaffected.

The new syntax is documented via "bikeshed issues-list -t".

Closes #1399