Preserve backward compat for link generation of datedStatuses
For datedStatuses, when Snapshot key has not been specified,
"This Version" and "Editor's Draft" would point to the same place.
Only include one of the two, and prefer "This Version" for compatibility
with the behavior prior to the introduction of the Snapshot key.

Side note: checking for "not 'version' in mac" is not strictly
necessary, given that when the status is one of datedStatuses,
'version' will always be in mac. But while that's true in current code,
future evolutions might change that, and better be safe than sorry.