Phrasing tweak in the documentation
As currently phrased, the "finally" part of the common steps can be a
bit confusing, as it may look like it tells you that the first step is
to do something in a virtual environment, which is confusing if you're
not using one, as instead of reading further and noticing it has
instructions for other environments, you want instead go back to read
earlier steps to figure out if you missed something that asked you to
set up one.

Rephrase to make it more obvious that this subsection is a list of
alternatives rather than a series of steps.