Update implementation-counting code
This change aligns the code with a recent change I made to the structure
of the https://github.com/w3c/mdn-spec-links JSON data.

The data change is that the value of the "engines" key previously was
just an integer but is now an array of zero or more engine names; e.g.,
["blink", "gecko", "safari"] (in the case of any feature that has
support in all three engines), ["gecko"] (in the case of a feature
that's only supported in gecko. Etc.