Support linking to steps.
Specifications often have algorithms with many (nested) ordered lists.
In those algorithms, steps often refer to other steps by number. But
those numbers currently have to be maintained manually and, when steps
are inserted, that can lead to a cascade of updated step number
references below. Worse, when two changes are in-flight to the same
algorithm, the step numbers are almost ensured to go wrong because
neither change need textually conflict with the other.

This change to bikeshed permits items in an ordered list to be given an
ID and referenced by name.

1. This is a traditional item.
1. {#myid} This is an item with an ID.
1. This items references [[#myid]].

The reference is turned into the string “step 2”. This also works for
nested steps where the reference will say something like “step 2.1.3”.