Style wpt elements to be minimally intrusive
Most readers of a spec are not here for the tests, so wpt elements,
particularly in their closed state, should get out of the way, and
disturb the general typography of the document as little as possible.
Hot pink wasn't the ideal color for that (and it's very close to the
w3c-chosen purple color used for candidate corrections anyway), so this
commit replaces that with gray. In the closed state, wpt elements are
made small, semi-transparent, right aligned, and fitting into the margin
between the elements they annotate.