Miscellaneous improvements for the WPT element (#2010)
* Use details element for WPT

Lists can be pretty long and get in the way of spec readability otherwise.

* Remove non standard attribute from output

The pathprefix attribute of the wpt element has done its job by the time
we generate the output, and it is a non-standard one, so remove it

* Handle 'hidden' on wpt elements

Document the fact that the hidden attribute on wpt elements lets you
exclude them from the output, while letting bikeshed check their
contents. And while we're at it, remove them entirely from the output.

Relates to https://github.com/tabatkins/bikeshed/issues/1895

* Don't inject wpt styles if "WPT Display" is none

* Enable using a title attribute on wpt elements to insert some commentary

Also, take advantage of that to add a description of wpt blocks in the
"Document conventions" sections of boilerplate files

* Make 'title' on wpt elements i18n-proof

* Merge neighboring wpt elements

Having multiple successive wpt elements in the source can happen due to
wanting a different pathprefix, but there's no reason to keep this
separation in the output since the prefixes have been processed away.

* Style wpt elements to be minimally intrusive

Most readers of a spec are not here for the tests, so wpt elements,
particularly in their closed state, should get out of the way, and
disturb the general typography of the document as little as possible.
Hot pink wasn't the ideal color for that (and it's very close to the
w3c-chosen purple color used for candidate corrections anyway), so this
commit replaces that with gray. In the closed state, wpt elements are
made small, semi-transparent, right aligned, and fitting into the margin
between the elements they annotate.

* Support more values for "WPT Display"

* Document the ability to merge wpt tags

* Turn on 'WPT Display: closed' for CSSWG EDs by default