Re-write status boilerplate for Media WG with new NOTE statuses
This update re-writes the boilerplate text for the Status of This Document
section of Media WG specifications, merging previous separate files into one
to convergence on the same text and ease maintenance, and adjusting paragraphs
for ED, FPWD, and WD of documents intended to become Notes.

The boilerplate text does not yet cover the PR and REC stages. To be adjusted
when group specifications get closer to these stages.