Support building multi-file specs from tar files
Adds a new --tar-file option which causes source files to be opened from
the tar file, instead of from the filesystem. It applies to the toplevel
source file and any file includes (`pre class=include`), so a tar file
for a multi-file spec can be built from a single file.

This makes it easy to do multi-file builds using the Bikeshed API
(http://api.csswg.org/bikeshed/). Bikeshed operates on the tar file in
memory, avoiding extracting arbitrary files to the local filesystem.
It supports whatever compression is supported by Python's `tarfile`,
e.g. gzip, bzip2, lzma.

Since this is meant for use with multi-file specs, input defaults to
`index.bs` or `Overview.bs` only, and doesn't search for other `.bs` or
`.src.html` files.

Output always defaults to `-` (stdout), because the input source file
doesn't exist in the filesystem and a proper output location can't be
guessed.

Usage:
- Output to a file:
  `bikeshed spec --tar-file=myspec.tar myindex.bs outfile.html`
- Output to stdout:
  `bikeshed spec --tar-file=myspec.tar myindex.bs -`, or
  `bikeshed spec --tar-file=myspec.tar myindex.bs`
- If using `index.bs` or `Overview.bs`:
  `bikeshed spec --tar-file=myspec.tar`, or
  `bikeshed spec myspec.tar`, or
  `bikeshed spec myspec.tar outfile.html`