Complete anchors/headings with data from Webref
This completes the update mechanism that produces the cross-references anchors
and headings to also use data from Webref. To keep changes minimal and avoid
introducing possibly conflicting anchors and headings here and there, data from
Webref is only used for specs that are not in Shepherd's database.

In practice, this adds definitions and headings from 267 specifications (~60 of
which only have headings), generating ~3.5MB of anchors data and ~5.2MB of
headings data.