Make webidl win over ecmascript, and infra win over i18n-glossary again
WebIDL creates proxy definitions of `EvalError`, `RangeError`, `ReferenceError`,
`SyntaxError`, `TypeError`, and `URIError` that are initially defined in
ECMAScript. The WebIDL definitions are typically preferred in web spec to
reference the concepts.

A previous commit made the Infra standard win over the Internationalization
glossary but that change was accidently rolled back by the following data
refresh:
https://github.com/tabatkins/bikeshed/commit/163bc60426b65f715b742c7901afb1a9815381b5