Support CDDL marking and linking
This makes Bikeshed process CDDL blocks à la `<pre class=cddl>` as described
in #2072 to:
- add highlighting (done by Pygments, the code merely sets the right class)
- wrap terms in `<dfn>` and `<a>` automatically, making it possible to
define them in, or reference them from, the rest of the prose.

Most of the logic is copied from the logic used to process IDL blocks.

As opposed to IDL definitions, CDDL definitions are not exported by default as
most CDDL definitions only apply to the underlying spec.

Support for CDDL definitions means 4 new definition types get introduced:
- `cddl-type`: roughly the equivalent of an IDL interface. Type definitions do
not have a `data-dfn-for` attribute.
- `cddl-key`: roughly the equivalent of an IDL attribute. Key definitions
always have a `data-dfn-for` attribute that links back to a CDDL type.
- `cddl-value`: roughly the equivalent of an enum-value in IDL. Value
definitions always have a `data-dfn-for` attribute that links back to a CDDL
type or to a CDDL key.
- `cddl-parameter`: used for generic parameter names (noting that no known spec
uses CDDL generics for the time being).

The definition types are prefixed with `cddl-` to avoid collision with other
types used for other purpose (e.g., `value` in CSS).

The code also collects CDDL definitions to produce a CDDL index at the end of
the spec. To accommodate specs like WebDriver BiDi that define two sets of
CDDL (remote end and local end), a mechanism gets added to associate CDDL
definitions with a given module:
1. The CDDL module must be defined with a `<dfn>` with a `data-cddl-module`
attribute set to a shortname for the CDDL module
2. CDDL blocks must add a `data-cddl-module` attribute set to a comma-separated
list of CDDL module shortnames they belong to. If a CDDL block does not have
that attribute, the code considers it is defined in all CDDL modules.

The index is split per module, using the `<dfn>` text as title for each module.

Note: Even when modules are used, CDDL definitions in a spec are part of the
same namespace, meaning a `foo` rule cannot be defined differently within a
single spec for two different CDDL modules.

CDDL parsing is done through a hand-made CDDL parser that follows the CDDL
grammar defined in RFC 8610, currently sitting under:
 https://github.com/tidoust/cddlparser

To ease authoring, a new shorthand notation gets introduced to reference CDDL:
`{^foo^}` is an autolink to a CDDL definition. FWIW, the shorthand was chosen
to mean "shortcut to CDDL code" on the grounds that:
- `{}` indicates a code block (for IDL)
- `^` means "cut" in CDDL

CDDL type definitions can become somewhat convoluted, the code does not
attempt to be too smart and won't autolink definitions that are too complex.