Improve support for generic parameters in CDDL blocks
The code now properly handles references to generic parameters within a type
definition, and does not get confused with types that have the same name and
that may be defined elsewhere.

It remains possible to end up with (insane) CDDL definitions such as:

```
T<T> = {
  T: T
}
```

The IDs are correctly set but there is no good way to disambuigate `{^T/T^}` in
such cases. Specs should not really need to reference generic parameters in
practice. This update simply disables generic `cddl` links and CDDL shorthand
syntax for now. In other words, the `T` generic parameter can only be
referenced through `<a cddl-parameter for=T>T</a>`, while `{^T/T^}` targets the
`T` key under the `T` type.

This could be improved later on if references to generic parameters turn out to
be useful in spec prose.