Warn/Report an error about duplicated CDDL keys/values
Code now detects when a CDDL block contains anonymous types that would create
duplicates. It dies when a key or a value is defined twice for the same CDDL
type, preventing any reference to the term. It warns when the duplicate is for
definitions that are not of the same type.

For example, an error is returned for:

```
err = [ "dupl", "dupl" ]
```

... while the following only returns an error because it remains possible to
reference individual terms by making the link type explicit:

```
warn = [ ( dupl: "val" ), "dupl" ]
```