Rebase CDDL tests on current version of spec-data/readonly
Data in `spec-data/readonly` needs to be refreshed for definitions of CDDL
types defined in the standard prelude in RFC 8610 to start becoming available
for cross-referencing purpose.

In the meantime, this update rebases the tests that include CDDL dfns. Build
typically reports link errors because of the unavailable RFC 8610 dfns. These
errors will disappear once data in `spec-data/readonly` gets refreshed and
CDDL tests rebased again.