Support CDDL marking and linking (#2977)
* Support CDDL marking and linking

This makes Bikeshed process CDDL blocks à la `<pre class=cddl>` as described
in #2072 to:
- add highlighting (done by Pygments, the code merely sets the right class)
- wrap terms in `<dfn>` and `<a>` automatically, making it possible to
define them in, or reference them from, the rest of the prose.

Most of the logic is copied from the logic used to process IDL blocks.

As opposed to IDL definitions, CDDL definitions are not exported by default as
most CDDL definitions only apply to the underlying spec.

Support for CDDL definitions means 4 new definition types get introduced:
- `cddl-type`: roughly the equivalent of an IDL interface. Type definitions do
not have a `data-dfn-for` attribute.
- `cddl-key`: roughly the equivalent of an IDL attribute. Key definitions
always have a `data-dfn-for` attribute that links back to a CDDL type.
- `cddl-value`: roughly the equivalent of an enum-value in IDL. Value
definitions always have a `data-dfn-for` attribute that links back to a CDDL
type or to a CDDL key.
- `cddl-parameter`: used for generic parameter names (noting that no known spec
uses CDDL generics for the time being).

The definition types are prefixed with `cddl-` to avoid collision with other
types used for other purpose (e.g., `value` in CSS).

The code also collects CDDL definitions to produce a CDDL index at the end of
the spec. To accommodate specs like WebDriver BiDi that define two sets of
CDDL (remote end and local end), a mechanism gets added to associate CDDL
definitions with a given module:
1. The CDDL module must be defined with a `<dfn>` with a `data-cddl-module`
attribute set to a shortname for the CDDL module
2. CDDL blocks must add a `data-cddl-module` attribute set to a comma-separated
list of CDDL module shortnames they belong to. If a CDDL block does not have
that attribute, the code considers it is defined in all CDDL modules.

The index is split per module, using the `<dfn>` text as title for each module.

Note: Even when modules are used, CDDL definitions in a spec are part of the
same namespace, meaning a `foo` rule cannot be defined differently within a
single spec for two different CDDL modules.

CDDL parsing is done through a hand-made CDDL parser that follows the CDDL
grammar defined in RFC 8610, currently sitting under:
 https://github.com/tidoust/cddlparser

To ease authoring, a new shorthand notation gets introduced to reference CDDL:
`{^foo^}` is an autolink to a CDDL definition. FWIW, the shorthand was chosen
to mean "shortcut to CDDL code" on the grounds that:
- `{}` indicates a code block (for IDL)
- `^` means "cut" in CDDL

CDDL type definitions can become somewhat convoluted, the code does not
attempt to be too smart and won't autolink definitions that are too complex.

* Add dependency to cddlparser Pypi package

* Drop now useless import statements

* Fix typing issues

`mypy` started reporting type issues as it can now look into the `cddlparser`
package. This update bumps the version of `cddlparser` to get better typing
info, and adds assertions when appropriate.

* Bump version of cddlparser to 0.3.0

Previous version was not fully compatible with Python 3.9 and Python 3.10. The
`cddlparser` project now runs tests under multiple versions of Python to
guarantee compatibility.

* Bump version of cddlparser to 0.4.0

New version returns more user-friendly error messages

* Detail CDDL support in documentation

* Fix typo in third CDDL test

The CDDL index contained the "local" module twice as a result.

* Support string/bareword refs for CDDL keys and values

Third test file also update to showcase and test more nested shorthand features
and ways to disambiguate apparent duplicates.

* Improve support for generic parameters in CDDL blocks

The code now properly handles references to generic parameters within a type
definition, and does not get confused with types that have the same name and
that may be defined elsewhere.

It remains possible to end up with (insane) CDDL definitions such as:

```
T<T> = {
  T: T
}
```

The IDs are correctly set but there is no good way to disambuigate `{^T/T^}` in
such cases. Specs should not really need to reference generic parameters in
practice. This update simply disables generic `cddl` links and CDDL shorthand
syntax for now. In other words, the `T` generic parameter can only be
referenced through `<a cddl-parameter for=T>T</a>`, while `{^T/T^}` targets the
`T` key under the `T` type.

This could be improved later on if references to generic parameters turn out to
be useful in spec prose.

* Warn/Report an error about duplicated CDDL keys/values

Code now detects when a CDDL block contains anonymous types that would create
duplicates. It dies when a key or a value is defined twice for the same CDDL
type, preventing any reference to the term. It warns when the duplicate is for
definitions that are not of the same type.

For example, an error is returned for:

```
err = [ "dupl", "dupl" ]
```

... while the following only returns an error because it remains possible to
reference individual terms by making the link type explicit:

```
warn = [ ( dupl: "val" ), "dupl" ]
```

* Rebase WebDriver BiDi test

* Define CDDL modules using a cddl-module dfn type

This create a new `cddl-module` dfn type to turn the beskope markup pattern
for defining CDDL modules into a more usual mechanism. The shorthand notation
can be used to reference CDDL modules as well.

* Move dupl detection logic to CDDL marker

That's a more natural place to put the logic and it simplifies things a bit as
well (no need to deal with additional linking texts).

* Auto-link CDDL prelude type refs to RFC 8610

Webref now contains the prelude types defined in RFC 8610, making it possible
to link to them automatically from other specs.

Note: The updated tests are correct... provided that `spec-data/readonly` gets
updated first through `bikeshed debug --refresh-data` (not done as part of this
commit to keep updates separated)

* Update note on "url" being used as a selector

* Lint code

* Rebase CDDL tests on current version of spec-data/readonly

Data in `spec-data/readonly` needs to be refreshed for definitions of CDDL
types defined in the standard prelude in RFC 8610 to start becoming available
for cross-referencing purpose.

In the meantime, this update rebases the tests that include CDDL dfns. Build
typically reports link errors because of the unavailable RFC 8610 dfns. These
errors will disappear once data in `spec-data/readonly` gets refreshed and
CDDL tests rebased again.

* Bump version of cddlparser

More to keep dependency up-to-date than for any other reason. Version 0.5.0
does not introduce any change or bug fix that affects its integration in
Bikeshed.

* Adjust CDDL link parsing to new parser

* Update bikeshed/h/parser/parser.py

* Adjust CDDL parsing code to new parsing functions

---------

Co-authored-by: Tab Atkins Jr. <jackalmage@gmail.com>